All issues addressed
