+++ golem_messages/serializer.py
@@ -1,13 +1,14 @@
-import cbor2
-import cbor2.types
 import collections
 import enum
 import functools
 import inspect
 import logging
-import pytz
 import sys
 import types
+
+import cbor2
+import cbor2.types
+import pytz
 
 logger = logging.getLogger('golem.core.simpleserializer')
 
@@ -21,9 +22,9 @@ def encode_message(encoder, value, fp):
     encoder.encode(cbor2.types.CBORTag(MESSAGE_TAG, serialized_msg), fp)
 
 
-def decode_message(decoder, value, fp, shareable_index):
-    from golem_messages import message
-    return message.Message.deserialize(
+def decode_message(decoder, value, fp, shareable_index):  # noqa pylint: disable=unused-argument
+    from golem_messages.message import base  # pylint: disable=cyclic-import
+    return base.Message.deserialize(
         value,
         decrypt_func=None,
         check_time=False
@@ -149,7 +150,7 @@ class DictCoder:
 
     @classmethod
     def _is_builtin(cls, obj):
-        if not type(obj) in cls.builtin_types:
+        if not type(obj) in cls.builtin_types:  # noqa This class will be refactored out in v1.6 pylint: disable=unidiomatic-typecheck
             return False
         return not isinstance(obj, types.InstanceType)
 
@@ -174,7 +175,7 @@ class CBORCoder(DictCoder):
 
 def encode_object(encoder, value, fp):
     if value is None:
-        return None
+        return
     obj_dict = CBORCoder.obj_to_dict(value)
     encoder.encode_semantic(
         OBJECT_TAG, obj_dict, fp,
@@ -182,7 +183,7 @@ def encode_object(encoder, value, fp):
     )
 
 
-def decode_object(decoder, value, fp, shareable_index=None):
+def decode_object(decoder, value, fp, shareable_index=None):  # noqa pylint: disable=unused-argument
     obj = CBORCoder.obj_from_dict(value)
     return obj
